refactor(admin): create utils component folder
