+++ src/admin/components/availability/utils/ProductAvailabilityItem.tsx
