+++ src/admin/components/availability/utils/ProductAvailabilityItemField.tsx
