+++ src/admin/components/availability/utils/ProductAvailabilityLoadingSkeleton.tsx
