+++ src/admin/components/availability/utils/ProductPicker.tsx
