+++ src/admin/components/availability/utils/ProductPickerItem.tsx
